+++ .github/workflows/test.yml
@@ -56,4 +56,4 @@ jobs:
       - name: Run custom theme checks
         working-directory: ./docs
         run: |
-          bundle exec ruby run_checks.rb
+          bundle exec ruby tests/test_html_pages.rb
