+++ docs/tests/custom_html_checks.rb
